requirements.txt
# OCR相关
mistral-ocr>=0.1.0
PyPDF2>=3.0.0
Pillow>=10.0.0

# 爬虫相关
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.0.0
webdriver-manager>=3.8.0

# 数据处理
markdown>=3.4.0
pyyaml>=6.0.0

# 工具
tqdm>=4.65.0
python-dotenv>=1.0.0
argparse>=1.4.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0 